Add files via upload
Basically, it's a project given at the University of Liberal Arts Bangladesh on the Algorithm lab course to find the shortest path to reach campus A to the main campus using any algorithm. I have used BFS algorithm as instructed and I also created another team's project which is slightly different.